school_saas/requirements/base.txt
# ── requirements/base.txt ──────────────────────────────────
django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=23.0
django-cors-headers>=4.0
celery>=5.3
redis>=5.0
django-celery-beat>=2.5
django-storages>=1.14
boto3>=1.34
Pillow>=10.0
psycopg2-binary>=2.9
python-decouple>=3.8
mailjet-rest>=1.3
drf-spectacular>=0.27
